requirements.txt
# Core dependencies
strands>=0.1.0
litellm>=1.0.0
python-dotenv>=1.0.0

# HTTP clients
httpx>=0.25.0
requests>=2.31.0

# Development dependencies
ruff>=0.1.0
pytest>=7.0.0
